R21.5 Link Colour + Fixes
+ Improved contrast on links.
- Added missing styling for progress inputs.
- Moved tag edit textarea slightly left.

I've been putting off changing the link colour for ages since I was afraid people would complain, but the old one was awfully low contrast and it needed to go. I was also tempted to change the colour on the header buttons, but feel that would have been too far. If it ain't broke, don't fix it as they say.